+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Updated Channel Name selector on channel page (selector should be future proof)
+
 ## [2.24.5] - 2026-08-18
 
 ### Fixed